external/webkit/LayoutTests/platform/android/webgl-conformance/dumpAsText/conformance/glsl/misc/shader-with-glcolor.vert-expected.txt

CONSOLE MESSAGE: line 24: *** Error compiling shader '[object WebGLShader]':ERROR: 0:5: 'gl_Color' : undeclared identifier 
ERROR: 0:5: 'assign' :  cannot convert from 'float' to 'Position highp 4-component vector of float'

vertex shader that uses gl_Color should fail

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

test: vertex shader that uses gl_Color should fail
test vertex shader
default fragment shaderPASS vertex shader that uses gl_Color should fail
PASS successfullyParsed is true

TEST COMPLETE
